Abstract factory method pattern confused??????????

+++ src/AbstractFactory/Packaging.java
@@ -1,5 +1,5 @@
 package AbstractFactory;
 
 public interface Packaging {
-	public float getCost();
+	public double getCost();
 }
